requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
idna==2.8
Jinja2==2.10
MarkupSafe==1.1.1
peewee==3.9.2
pycparser==2.19
pytafseer==0.1.1
python-dotenv==0.10.1
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
